ipq40xx: ASoC: Change descriptor tracking logic
Earlier, we relied on the h/w flipping the 'own' bit of the
descriptor to identify the i/o completed buffers. Hence
driver marked all the descriptors as h/w owned by default.
This resulted in some unwanted data too being played back.
To avoid this, track buffer usage using read/write pointers
and give buffers to h/w as and when they need to be played.

Change-Id: Ic763c8d545d0b0b0adb36a5c2a12bd948c40a105
Signed-off-by: Natesh babu <nrameg@codeaurora.org>