ipq40xx: Fix to handle audio pop noise during bootup
When the sound card registration fails, the audio TLMM change is also
reverted. Due to this, the pins are seen to toggle causing pop noise
during kernel bootup. To avoid this, the pins are set to GPIO state and
moved to audio functionality only when the sound card registration is
successful.

Change-Id: I70bb3968250e3df7c68bfa2c5f8d188d04848671
Signed-off-by: G Dhivya <gdhivya@codeaurora.org>