ipq4019: Adding support for Winbond W25M02GV SPI NAND
W25M02GV consists of two stacked dies each of 128MB.
Only a single die can be active at a time and hence
we require a die select command to be given along with
the die ID (0 or 1)  before read, write and erase
operations can be done in a particular die. The changes
for supporting die select command has been implemented
in the driver.

Change-Id: I8e0de6d9e509b7d7c2087dd979b50b0d6f8e45a6
Signed-off-by: G Dhivya <gdhivya@codeaurora.org>