bridge: Fix dropping of Router Advertisement packets
'proxyarp_replied' for IPV4 seems to hit for IPV6 Router
Advertisement packets and eventually dropping it, resulting
in the failure to provide Global IPV6 address for the
reference station. Fix this by having separate variables to
track Proxy ARP operation for IPV4 and IPV6. Also drop frames
only if proxyarp_replied/ proxyndisc_replied is set to true
and its an ARP packet

Change-Id: I4aa592af0954a423fd268b3b5a5f44a3c13e9ca2
Signed-off-by: Zhu Ken <guigenz@codeaurora.org>
Signed-off-by: Mohammed Shafi Shajakhan <mohammed@codeaurora.org>