[skb recycler] Don't call BUG in the notifier
The notifier was meant to be terminal, i.e. after propagating the
event it stops the system.
This is causing confusion however, under the assumption that there is
an actual bug in the notifier code.

This commit removes the BUG call in the notifier, so that the event
originator gets to stop the system (as it was done before the notifier
was introduced).

Change-Id: I5a7af0374dc4991539d742712331de03aa3833d1
Signed-off-by: Cristian Prundeanu <cprundea@codeaurora.org>