CHROMIUM: mac80211: mesh path selection stability enhancements
To increase path selection stability, we want to favor
the same nexthop and only switch when the alternative
is considerablly better than the current one. The new
logic is as follows:

0. throw away updates from past epochs

1. always update info if the nexthop is the same, because
  - We want to get latest link metric for the chosen nexthop
  - This will update lifetime of the path so we can keep using
    the same nexthop

2. if update is for a different nexthop, update if the new
   link metric is path_switch_threshold % better.
  - path_switch_threshold is tunable from debugfs

Note that the existing implementation will always update
if it sees a new epoch, but here we want to wait for
updates for the same nexthop. This may cause the path to
expire if we don't see the same nexthop update in time.
This can be helped by doubling the path lifetime (currently
6s) so we can afford to lose 1 update for the same nexthop.

BUG=b:26546424
TEST=manually on 4 WWs + 1 AK forming a mesh

Change-Id: Ic1074c326f10724baa3880ffd36d8667bebe2de9
Reviewed-on: https://chromium-review.googlesource.com/322776
Commit-Ready: Julan Hsu <julanhsu@google.com>
Tested-by: Julan Hsu <julanhsu@google.com>
Signed-off-by: Julan Hsu <julanhsu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/340404
Commit-Ready: Yixiang Li <yixiang@google.com>
Tested-by: Yixiang Li <yixiang@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Signed-off-by: Tamizh chelvam <tamizhchelvam@codeaurora.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>