CHROMIUM: mesh: do not delete expired path if it proxies for itself
An active path sometimes expires due to not getting updated. In this
case, mesh code checks if it can find a proxy for this destination.
If yes, it deletes the path assuming the path has become obsolete.
However, in certain use cases a destination can be a proxy for itself,
and it does not make sense to delete the path as this is still the
path which the packet will go through. The change is to add a check
to not delete the path if the proxy and the destination are the same.

BUG=b:28380230
TEST=pump traffic on a mesh composed of 3 whirlwinds. observe path
expiration status and no flapping of path entry overnight.

Reviewed-on: https://chromium-review.googlesource.com/346224
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>

Change-Id: I4c922b994f9b74921414bc712b4bb636f9415630
Signed-off-by: Julan Hsu <julanhsu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/347304
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>