skbuff/mac80211: introduce and use skb_put_zero()
This pattern was introduced a number of times in mac80211 just now,
and since it's present in a number of other places it makes sense
to add a little helper for it.

This just adds the helper and transforms the mac80211 code, a later
patch will transform other places.

Acked-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit e45a79da863c199d7c47b1ee6d33cee23c89eac1)

Change-Id: Iacb1ae4c34565660ad09f3c36ade239c43fd14de
Signed-off-by: Gokul Sriram Palanisamy <gokulsri@codeaurora.org>