ASoC: wcd934x: fix master clock divider settings
Currently the master clock always gets divided by 2 in the codec
driver. This is not correct as it is possible the external clock
supplied is already divided. Fix this by removing the clock
divider register write from resource manager driver and update
the defaults for I2C mode to make sure clock divider is set to 0.

Change-Id: I8667ff5e167810a133b55022231c3bc224afc46e
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>