ipq40xx: spi_qsd: Do not connect to sps bam pipe for fifo transfer
spi_qsd driver is connecting to sps bam pipe for each spi transfer
when the dma is enabled in dts. But to select bam for a particular
transfer more conditions need to be met. Connecting to bam pipe
for each transfer utilizing more CPU.

In order overcome this bam pipe connect is used only when all
conditions for a bam transfer is met.

Change-Id: I50bb0b247596752230b1807615f12fcc83172380
Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>