netfilter: Do not cache a dst if fake rtbale flag is set.
If the bridge netfilter iptables calls are enabled,
the fake rtable is dropped from a skb in the
br_dev_queue_push_xmit function, when the skb
(a broadcast packet) is flooded to the ports. But if the
bridge netfilter iptables calls are disabled after the dst
is cached, since the skb doesn't go through the forwarding
path anymore, the fake rtable is not dropped. After the flood,
this skb does through the pre-routing hook in the ip_rcv function.
The fake rtable's input function is NULL and it causes crash
in the ip_rcv_finish function.

Change-Id: I20fbe8badc304d9244c31ad0f5d0f3271300dcdc
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>