CHROMIUM: mesh: cap link metric max weight of new sample to 5%
This will make sure that new link metric formula always gives same or
less weight to new samples compared to the baseline formula, which uses
a static weight of 5%. In other words, with this CL, path changes
should be about the same or less frequent in all traffic scenarios.

BUG=b:64971145
TEST=manually on 3 gales, tested under no active traffic load as well as
under iperf traffic between 2.

Change-Id: Id0ac782d3cbe8747d27daba74147a16de553c536
Signed-off-by: Narayanraddi Masti <nmasti@chromium.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>