msm_serial_hs: Change the wakeup interrupt enable  sequence
The wakeup variable keeps track of the 2 edge interrupts needed to detect
wakeup from deep sleep, when the second edge interrupt is detected the
wake byte is injected into the tty buffer.
On some boards due to noisy GPIO pins an interrupt is detected as soon as
the wakeup irq is enabled and before the variable is reset leading to
bogus wakeup bytes being injected into the serial buffer.
To protect against this, make sure the wakeup interrupt is enabled after
all the state variables are reset to protect against spurious wakeups.
Brought the change from msm-3.18 and manually merged it in 3.14,from
commit :747125be3ec7e2db59c852fa1385548922de1702

Change-Id: I8002a67d8d9cb41709049ff34f35717f3dd0acd0
Acked-by: Yijiang Yuan <yijiangy@qti.qualcomm.com>
Signed-off-by: Girish Mahadevan <girishm@codeaurora.org>
Signed-off-by: sameeruddin shaik <samesh@codeaurora.org>