wil6210: fix call to wil6210_disconnect during unload
Move the call to wil6210_disconnect so it will be called
before unregister_netdevice. This is because it calls
netif_carrier_off which is forbidden to call on an
unregistered net device. Calling netif_carrier_off can
add a link watch event which might be handled after
net device was freed, causing a kernel oops.

Change-Id: Ia9ccf2a9734a0cbd0a961b4873e36439ae9a928e
Signed-off-by: Lior David <liord@codeaurora.org>