clk: add dynamic debug prints
In some of the silent reboot cases, it is found that root cause is due to
clock related operations. In order to improve the debugging in such cases
added the dynamic debug prints for the following clock operations.

	1) clk_enable
	2) clk_disable
	3) clk_set_parent
	3) clk_set_rate

These dynamic debug prints can be enabled through either of the following
options

1) append the following query to bootargs
  "dyndbg=\\"file drivers/clk/clk.c +p\\""

or

2) by writing into the dynamic debug control file as
  echo -n 'file drivers/clk/clk.c +p' > <debugfs>/dynamic_debug/control

By default, these prints will be disabled

Change-Id: I6c103f7dbd37222e0fab407a46cf3c1dacd2bfaf
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>