clk: qcom: ipq8074: use actual clock and reset defines
Currently dummy clock is mapped with DUMMY_CLK. Instead of this,
we can use actual clock and reset defines and map the clocks to
dummy clock structure. This will help in

1. Bring up where all the changes needs to be done in clock
   driver alone
2. The existing user of v1 does not have to change clock and
   reset mapping for dummy clocks. It can use the proper clock and
   reset names in dtsi.

Also, remove the dummy clock file and maintain single driver file
for ipq8074 v1 and ipq8074 v2 with different compatible strings.

Change-Id: Ib11b7956dd7758eb65fd0350bf45bfc2b430296e
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>