wil6210: increase RX status ring size
HW can halt in case it has available descriptors but has no RX status
messages for storing their completions.
Increase RX status ring size in order to prevent such a case.

In addition, add a module param to allow changing the count of
RX buffer IDs. The count is checked in order to guarantee that
it doesn't exceed the RX status ring size.

Change-Id: Iea134dc68451816babbd9a40df7a6684bc947468
Signed-off-by: Maya Erez <merez@codeaurora.org>