qseecom: ipq: Enabling Services in sample app
This patch involves changes which enables tz
logging in ipq807x and ipq40xx platform.

TZ logging is not enabled in ipq806x platform.

This patch also involves changes which allows
the user to load libraries required for sample
app in ipq807x and ipq40xx platform.

Libraries are in-built in ipq806x platform.

Change-Id: I2a62befa15ea7f2fde5c2803d1afdf7b7725d263
Signed-off-by: speriaka <speriaka@codeaurora.org>