cnss2: update qmi memory request processing based on the new qmi headers
The new qmi flow will pass a memory request for multiple segments in
one single message and host will respond back with memory address
and size for all segements in one single message. Modify the code
to inspect each memory segment request, process and provide the memory
region for it.

Change-Id: If3080f1bc542cb6ef6e0e1eb0c89a6344ac661af
Signed-off-by: Balamurugan Mahalingam <bmahalin@codeaurora.org>