FROMLIST: PM / devfreq: Add support for QCOM devfreq firmware
The firmware present in some QCOM chipsets offloads the steps necessary for
changing the frequency of some devices (Eg: L3). This driver implements the
devfreq interface for this firmware so that various governors could be used
to scale the frequency of these devices.

Each client (say cluster 0 and cluster 1) that wants to vote for a
particular device's frequency (say, L3 frequency) is represented as a
separate voter device (qcom,devfreq-fw-voter) that's a child of the
firmware device (qcom,devfreq-fw).

[sibis: Fix unable to allocate ftbl-base]

PM / devfreq: Generic CPU frequency to device frequency mapping governor

Many CPU architectures have caches that can scale independent of the CPUs.
Frequency scaling of the caches is necessary to make sure the cache is not
a performance bottleneck that leads to poor performance and power. The same
idea applies for RAM/DDR.

To achieve this, this patch adds a generic devfreq governor that takes the
current frequency of each CPU frequency domain and then adjusts the
frequency of the cache (or any devfreq device) based on the frequency of
the CPUs. It listens to CPU frequency transition notifiers to keep itself
up to date on the current CPU frequency.

To decide the frequency of the device, the governor does one of the
following:

* Uses a CPU frequency to device frequency mapping table
  - Either one mapping table used for all CPU freq policies (typically used
    for system with homogeneous cores/clusters that have the same OPPs).
  - One mapping table per CPU freq policy (typically used for ASMP systems
    with heterogeneous CPUs with different OPPs)

OR

* Scales the device frequency in proportion to the CPU frequency. So, if
  the CPUs are running at their max frequency, the device runs at its max
  frequency.  If the CPUs are running at their min frequency, the device
  runs at its min frequency. And interpolated for frequencies in between.

INTERNAL: arm64: dts: Enable L3 devfreq and cpufreq-map governor

Enable L3 devfreq node and cpufreq-map governor node for SDM845
SoCs

CHROMIUM: config: Enable L3 Devfreq/cpufreq-map governor configs

This is equivalent of:

echo 'CONFIG_DEVFREQ_GOV_CPUFREQ_MAP=y' >> \
     chromeos/config/arm64/chromiumos-arm64.flavour.config
    echo 'CONFIG_ARM_QCOM_DEVFREQ_FW=y' >> \
     chromeos/config/arm64/chromiumos-arm64.flavour.config
    echo 'CONFIG_DEVFREQ_GOV_CPUFREQ_MAP=y' >> \
     chromeos/config/arm64/chromiumos-qualcomm.flavour.config
    echo 'CONFIG_ARM_QCOM_DEVFREQ_FW=y' >> \
     chromeos/config/arm64/chromiumos-qualcomm.flavour.config
./chromeos/scripts/kernelconfig olddefconfig

Change-Id: Id580c428dd742b1ad3b6ab49a3eec6f5545937f8
Signed-off-by: Sibi Sankar <sibis@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>