FROMLIST: clk: qcom: Add lpass clock controller driver for SDM845
Add support for the lpass clock controller found on SDM845 based devices.
This would allow lpass peripheral loader drivers to control the clocks to
bring the subsystem out of reset.
LPASS clocks present on the global clock controller would be registered
with the clock framework based on the device tree flag.

BUG=b:112880092
TEST=Build and boot

Change-Id: I1328c4b73410b4ebfe3be2f274b9c8b852c82bc1
Signed-off-by: Taniya Das <tdas@codeaurora.org>
Signed-off-by: Mohammed Thasleem <mthasl@codeaurora.org>