ipq: sec-upgrade: return the proper error codes
when kstrtol fails, we are returning zero, so that shell assumes that
no data has been written, and it is writing continuously which results
in a infinite loop.

This patch fixes the same by returning the proper error value and takes
care of the similar issue in other places as well.

Change-Id: I86d26564779d43e62e6380335f889ba459742fb2
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>