ipq: sec-upgrade: return the proper error codes
when kstrtol fails, we are returning zero, so that shell assumes that
no data has been written, and it is writing continuously which results
in a infinite loop.

This patch fixes the same by returning the proper error value and takes
care of the similar issue in other places as well.

Change-Id: Icdb8d32bed1fb13ae0075a5806f71a9d30587f01
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>