net: timestamp: fix the performance issue
The funciton ptp_classify_raw will consume about 13% utilization of CPU
in the performance test.

Adding a feature bit to indicate whether the packet needs to be inspected
for timestamping in the PHY driver to avoid the unnecessary BPF program
running in the RX path.

The added bit ADVERTISED_PTP should be set in the PHY driver API
phy_driver->hwtstamp when the timestamp function of PHY device is enabled.

Change-Id: I4831bca31a0b3fc68ecef18af5eaaac8a10ac72b
Signed-off-by: Luo Jie <luoj@codeaurora.org>