dcc: Added a dcc magic and timestamp to dump list
A unique magic number is required in the dcc dump
to identify the start of a dump entry in the etb
which is a circular buffer.

Timestamp is required to identfiy the dump time.

The parser shall expect a 32-bit dcc magic as the
first entry and gcnt_lo and gcnt_hi as the next
two successive entries and then followed by the
values of the regions configured.

On config_reset, First two configuration entries,
corresponding to dcc_magic and gcnt entries will
not be cleared.

Change-Id: I400837dbc067ade9f5f322e382b273cb5e1d4de8
Signed-off-by: Gokul Sriram Palanisamy <gokulsri@codeaurora.org>