CHROMIUM: ath10k: Fill the channel survey results for WCN3990 correctly
The host driver currently expects the channel info event to be
received in pairs for all the channels, i.e. the first
chan_info event for a particular channel will not have the
COMPLETE flag set and the second chan_info event for the
same channel will have the COMPLETE flag set.

The HL2.0 firmware sends only one channel info event per channel
which is scanned without the COMPLETE flag set. After sending the
chan_info_event for all the channels, the HL2.0 firmware sends a
chan_info_event with COMPLETE flag set to indicate the completion
of the channel info event.

The firmware does not indicate this behavior with any service bitmap
and hence a new firmware feature flag is used to handle the modified
parsing of the channel info events, in the host driver, for the
firmware which sends single channel info event per scanned channel.

Tested HW: WCN3990
Tested FW: WLAN.HL.2.0-01188-QCAHLSWMTPLZ-1

BUG=b:112464483
TEST=wifi_matfunc on QCA6174A; basic sanity on WCN3990

Change-Id: Ifa80c272ecccabbaaeea253f79a3ff497b4b3b5c
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>
Signed-off-by: Mohammed Thasleem <mthasl@codeaurora.org>