INTERNAL: drm/msm/A6xx: Add devfreq support for A6xx
Implement routines to estimate GPU busy time and fetching the
current frequency for the polling interval. This is required by
the devfreq framework which recommends a frequency change if needed.
The driver code then tries to set this new frequency on the GPU by
sending an Out Of Band(OOB) request to the GMU.

Change-Id: I5110940c30a23295afe79f60ec41907c3ec8b58d
Signed-off-by: Sharat Masetty <smasetty@codeaurora.org>
Signed-off-by: Mohammed Thasleem <mthasl@codeaurora.org>