INTERNAL: drm/msm/A6xx: Add devfreq support for A6xx
Implement routines to estimate GPU busy time and fetching the
current frequency for the polling interval. This is required by
the devfreq framework which recommends a frequency change if needed.
The driver code then tries to set this new frequency on the GPU by
sending an Out Of Band(OOB) request to the GMU

Change-Id: I92c5564170c89ad1b5f2b2789c1ebfe43feba80f
Signed-off-by: Sharat Masetty <smasetty@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>