ipq806x: mmci: Timeout using s/w timer,when there are no h/w irqs
When we write in the write protected region of MTFC32GJVED-4M IT
part, we are getting data timeout error at first. After DMA being
reset, when another write job is submitted, we are not getting any
irq's from dma as well as sdcc controller. Therefore, introduced
software timer to get out of this situation.

Change-Id: I6c0d748f31d2c8997a7875ba2f9b39086183ff64
Signed-off-by: Balaji Prakash Jagadeesan <bjagadee@codeaurora.org>