FROMLIST: timer: Export next wakeup time of a CPU
Knowing the sleep duration of CPUs, is known to be needed while selecting
the most energy efficient idle state for a CPU or a group of CPUs.

However, to be able to compute the sleep duration, we need to know at what
time the next expected wakeup is for the CPU. Therefore, let's export this
information via a new function, tick_nohz_get_next_wakeup(). Following
changes make use of it.

Cc: Lina Iyer <ilina@codeaurora.org>

BUG=None
TEST=Build and boot

Signed-off-by: Raju P.L.S.S.S.N <rplsssn@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10478021/)

Change-Id: Ic7e448d75bf5a1454a64887b36a6171cc18c39be
Signed-off-by: Mohammed Thasleem <mthasl@codeaurora.org>