sound: qcom: Rename pinmux used for Audio master mode
This patch is necessary as it renames the pinmux
used in master mode to its older pinmux name which
other platform refers to.

Without this change, pinctrl state will not be found
in other platforms as they will be looking for older
pinctrl named state, because of which we might face
audio driver initialization issues and so the devices
might not get probed.

Change-Id: I7efe89cb14984afcbace34b0a455208e5092e17e
Signed-off-by: speriaka <speriaka@codeaurora.org>