INTERNAL: scsi: qcom-ufs: Add support for bus voting using ICB framework
Adapt to the new ICB framework for bus bandwidth voting.

This requires the source/destination port ids.
Also this requires a tuple of values.

The tuple is for two different paths - from UFS master
to BIMC slave. The other is from CPU master to UFS slave.
This tuple consists of the average and peak bandwidth.

Change-Id: I12b12939b5825d45cf224a5c5d217099fcb7322a
CRs-fixed: 2310101
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>