INTERNAL: scsi: qcom-ufs: Add support for bus voting using ICB framework
Adapt to the new ICB framework for bus bandwidth voting.

This requires the source/destination port ids.
Also this requires a tuple of values.

The tuple is for two different paths - from UFS master
to BIMC slave. The other is from CPU master to UFS slave.
This tuple consists of the average and peak bandwidth.

CRs-fixed: 2310101
Change-Id: Ie9de62ab00ff02e06a08fbb3bfab2cab9e49f9fe
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>