INTERNAL: scsi: qcom-ufs: Add support for bus voting using ICB framework
Adapt to the new ICB framework for bus bandwidth voting.

This requires the source/destination port ids.
Also this requires a tuple of values.

The tuple is for two different paths - from UFS master
to BIMC slave. The other is from CPU master to UFS slave.
This tuple consists of the average and peak bandwidth.

CRs-fixed: 2310101

Change-Id: I52c0e685bb5079b6fcf12a212b9dfae7bb845eb3
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>
Signed-off-by: Mohammed Thasleem <mthasl@codeaurora.org>