INTERNAL: drm/msm/A6xx: Send the right perf index value to GMU.
The index of the perf table was being set in the wrong bit position
in the register. With this fix, the GPU clock can be seen running at
desired frequency.
INTERNAL: drm/msm/a6xx: Add inactive_period for a6xx
The target definition for a630 didn't set a reasonable
value for inactive_period so it defaulted to zero and
we were essentially powering down after every submission.
Set it back to the default value to keep the GPU from
bouncing too much during regular workloads.

Change-Id: I913513f334b3f699e21d55abe00a029dbbcd9b20
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>