FROMLIST: ath10k: Enable Factory Test Mode for WCN3990
The support to put WCN3990 firmware into Factory
test mode is not present currently. The WCN3990
firmware can operate in Factory test mode based
on the mode it receives in the wlan enable message
from the host driver.

When the host driver is started in testmode send
the operating mode as UTF mode, to the WCN3990
firmware, in the wlan enable message to start the
firmware in Factory test mode.

Tested on: WCN3990
Tested FW: WLAN.HL.2.0-01192-QCAHLSWMTPLZ-1.

BUG=b:78530936
TEST=wifi_matfunc on QCA6174A; basic sanity on WCN3990

Change-Id: Ia20f853dcacbce255ac84a62522f42aa8a7b7713
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>
(am from https://patchwork.kernel.org/patch/10579585/)

Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>