CHROMIUM: mac80211: mesh airtime link metric fixes
Airtime link metric value is a function of bit rate
(in kbps) and paket error rate, and good estimates
of them are important to derive meaningful metric and
to attain desirable routing decisions.

This CL aims to improve exisiting implementation which
uses instantaneous bitrate value and does not account for
retries when calculating packet error rate. It maintains
a exponential moving average for bitrate, and the weight
is tunable from debugfs. It also includes retry count for
the purpose of calculating packet error rate. For example,
if a packet is tried 3 times and the 4th attempt succeeds,
then the packet error rate is 0.75.

BUG=b:26546424
TEST=manually on 4 WWs and 1 AK forming a mesh

Change-Id: I78d1300a5ab794b13dbc08c7ac83ad6e33ef67d1
Signed-off-by: Julan Hsu <julanhsu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/340402
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Signed-off-by: Tamizh chelvam <tamizhchelvam@codeaurora.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>