INTERNAL: drm/msm/A6xx: Add devfreq support for A6xx
Implement routines to estimate GPU busy time and fetching the
current frequency for the polling interval. This is required by
the devfreq framework which recommends a frequency change if needed.
The driver code then tries to set this new frequency on the GPU by
sending an Out Of Band(OOB) request to the GMU

Change-Id: If924aac28c269da6b0be7268ec6b346e4d5b5216
Signed-off-by: Sharat Masetty <smasetty@codeaurora.org>
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>
Signed-off-by: Amit Kushwaha <kushwaha@codeaurora.org>
Signed-off-by: Mohammed Thasleem <mthasl@codeaurora.org>
Signed-off-by: Parag Badge <pbadge@codeaurora.org>