FROMLIST:ath10k: Add support to configure TID specific config
This patch add ops for set_tid_conf to support
station specific retry and aggregation configuration for a TID.
station and configuration parameter(TID, configuration changed
parameter through ieee80211_tid_conf) information will be passed
as an argument for this ops.
Since target accepts one parameter as a retry count, ath10k
pass retry_long along with TID and station's mac address to
target in the command and store this if TID_RETRY_CONF_CHANGED.
And pass enable/disable aggregation for the TID
if TID_AGGR_CONF_CHANGED is passed. Depends on the changed
parameter the respective value stored in arsta struct.

Suppose if the station entry is not mentioned in the command,
the configuration will be applied for all the connected stations
in the vif.

And this configuration will be stored in arvif to apply this
configuration for newly connecting station.

Testing:
        * Tested HW: QCA9888
        * Tested FW: 10.4-3.5.1-00052

(Applied from https://patchwork.kernel.org/patch/10652337/)

Change-Id: I837ffb8c891ea7315c39049489651f490f89127a
Signed-off-by: Tamizh chelvam <tamizhr@codeaurora.org>
Signed-off-by: Vikas Patel <vikpatel@codeaurora.org>