Revert "ipq807x: diag: Allow only PERIPHERAL_WCNSS."
Commit d60481335fa8703ffd356822333d004c75d122c2 ("ipq807x: diag:
Allow only PERIPHERAL_WCNSS.") allows the various update messages
to be sent only to the PERIPHERAL_WCNSS. The updates will be sent
to the peripherals, whose control channel is open, other than the
currently connected peripheral.

The control channel for a peripheral will be open only when diag
client socket of APPS is connected to the diag service of that
peripheral. Reverting this change so that whichever peripheral
is connected, the updates will be sent to all other peripherals.

Signed-off-by: Balaji Prakash Jagadeesan <bjagadee@codeaurora.org>
Change-Id: Iaf511bb7701b0347ffa83ebe00647b90da42f7e5