FROMGIT: regulator: qcom-smd: Batch up requests for disabled regulators
In some scenarios the early stages of the boot chain has configured
regulators to be in a required state, but the later stages has skipped
to inform the RPM about it's requirements.

But as the SMD RPM regulators are being initialized voltage change
requests will be issued to align the voltage with the valid ranges. The
RPM aggregates all parameters for the specific regulator, the voltage
will be adjusted and the "enabled" state will be "off" - and the
regulator is turned off.

This patch addresses this problem by caching the requested enable state,
voltage and load and send the parameters in a batch, depending on the
enable state - effectively delaying the voltage request for disabled
regulators.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit fd805d99352c89b5cc1ca62ca5644840c9f02f07
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/regulator.git/commit/?h=for-next)
Signed-off-by: Khasim Syed Mohammed <khasim.mohammed@linaro.org>

Change-Id: I1cffa19a6ac6b8e4306b1d3c2a3a8191e1114e81
Signed-off-by: Vamsi Singamsetty <vamssi@codeaurora.org>