ipq: rproc: Update ADSP crashdump device structure
Crashdump device name for both ADSP and WCSS is "q6mem".
Renaming the devices so that, when both subsystems are
crashed at the same time, there won't be any clash in
the device name. Also, update ss_name entry to get the
correct reserved memory node("lpass") as in dts.

Change-Id: I4db421228f7834671b258f612a876bd37c1886b9
Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>