INTERNAL: cfg80211: Add support to limit number of beacon intervals
This change allows the host drivers to specify the
maximum number of unique beacon intervals which it
supports for AP interfaces.
After the max limit for different beacon intervals
is reached, all the subsequent creation of AP interfaces,
with beacon intervals different from the ones already running,
will fail.

If the driver does not specify this limit, then the
old behavior, where the AP interfaces can be brought
up with any beacon interval, is retained.

CRs-Fixed: 2398733
Change-Id: I20c219d543bf2d6e67b73e4d8c2462635f35a6a3
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>