INTERNAL: cfg80211: Fix check for num unique bi
The number of unique beacon interval check is
done unconditionally for all wiphy without
checking if the driver has exposed any such
limitations on number of unique beacon intervals.

The limit for number of unique beacon intervals
is to be checked only if the driver has set the
parameter in the interface combinations, else
any number of beacon intervals should be allowed.

Fix the checking for maximum number of beacon intervals.

Change-Id: I602ad8889da2109604c8fa9e0b2f341370bb982e
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>